// src/data/evolutions.gen.ts
export const EVOLUTION_METHODS = {
  EVO_FRIENDSHIP: 1,
  EVO_FRIENDSHIP_DAY: 2,
  EVO_FRIENDSHIP_NIGHT: 3,
  EVO_LEVEL: 4,
  EVO_TRADE: 5,
  EVO_TRADE_ITEM: 6,
  EVO_ITEM: 7,
  EVO_LEVEL_ATK_GT_DEF: 8,
  EVO_LEVEL_ATK_EQ_DEF: 9,
  EVO_LEVEL_ATK_LT_DEF: 10,
  EVO_LEVEL_SILCOON: 11,
  EVO_LEVEL_CASCOON: 12,
  EVO_LEVEL_NINJASK: 13,
  EVO_LEVEL_SHEDINJA: 14,
  EVO_BEAUTY: 15,
} as const;

export const EVOLUTION_MODES = {
  EVO_MODE_NORMAL: 0,
  EVO_MODE_TRADE: 1,
  EVO_MODE_ITEM_USE: 2,
  EVO_MODE_ITEM_CHECK: 3,
} as const;

export interface EvolutionRule {
  method: number;
  param: number;
  targetSpecies: number;
}